requirements.txt
aiosmtpd~=1.4.6
pydantic~=2.11.7
PyYAML~=6.0.2
alibabacloud-tea-openapi==0.4.1
alibabacloud_dm20151123==1.6.1